WIP Security Hardening Pt.2
Not sure what the right way is to add a new custom field to Payment Requests as part of Stripe Plus so just manually editing the json files for now. The idea is to generate and store a random access code that will be included in the request URL in the future to prevent payment URLs from easily guessed.